Elevator floor buttons functionality added
In the PRESS command added the ability to press elevator floor buttons and to change floors

Renamed command back to commands.
GO command no prints the description of the new location of the player when they move.
Renamed createTargetRoom to getTargetRoom.
Added getCurrentFLoorNumber() method to the Elevator class, returns 1 if on the first floor etc..
Added different text if an exit leads to an elevator in the LOOK command.
Added contents vector with addObject and removeObject methods to the location class. This vector will be used to store all the physical objects that are in the location.
Fixed a bug where isNumber(int) was always returning false